+++ lib/src/models/status.dart
@@ -1,12 +1,15 @@
 import 'package:equatable/equatable.dart';
 
+import '../managers/room.dart';
 import '../utils/field_value.dart';
 
-final class StatusKeys {
-  const StatusKeys._();
+class StatusKeys {
+  final String isOnline;
+  final String lastSeen;
 
-  static const isOnline = 'isOnline';
-  static const lastSeen = 'lastSeen';
+  static StatusKeys get i => RoomManager.i.modelConfigs.statusKeys;
+
+  const StatusKeys({this.isOnline = 'isOnline', this.lastSeen = 'lastSeen'});
 }
 
 class Status extends Equatable {
@@ -22,8 +25,9 @@ class Status extends Equatable {
   factory Status.parse(Object? source) {
     if (source is Status) return source;
     if (source is! Map) return Status.empty();
-    final isOnline = source[StatusKeys.isOnline];
-    final lastSeen = source[StatusKeys.lastSeen];
+    final keys = StatusKeys.i;
+    final isOnline = source[keys.isOnline];
+    final lastSeen = source[keys.lastSeen];
     return Status._(
       isOnline: isOnline is bool ? isOnline : false,
       lastSeen: ChatValueTimestamp.parse(lastSeen),
@@ -31,9 +35,10 @@ class Status extends Equatable {
   }
 
   Map<String, dynamic> get source {
+    final keys = StatusKeys.i;
     return {
-      if (isOnline) StatusKeys.isOnline: isOnline,
-      if (!lastSeen.isEmpty) StatusKeys.lastSeen: lastSeen,
+      if (isOnline) keys.isOnline: isOnline,
+      if (!lastSeen.isEmpty) keys.lastSeen: lastSeen,
     };
   }